Backend

+++ Backend/Controllers/RecipeManagerController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Backend.Models.BindingModels;
 using Backend.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Backend.Controllers
 {
@@ -21,12 +22,15 @@ namespace Backend.Controllers
         /// <response code="400">Некорректные значения</response>
         /// <response code="500">Внутренняя ошибка (читать сообщение в ответе)</response>
         [HttpPost("api/[controller]/add")]
+        [Authorize(Roles = "admin, open")]
         public IActionResult Post(FullInfoRecipeBindingModel recipe)
         {
             try
             {
                 if (recipe == null)
                     return BadRequest();
+
+                recipe.author = User.Identity.Name;
 
                 if(AddIntoDB(recipe))
                 {
@@ -69,7 +73,17 @@ namespace Backend.Controllers
                         {
                             IngredientList i = new IngredientList();
                             i.Idrecipe = curidrecipe;
-                            i.Idingredient = ingredient.idIngredientList;
+                            i.Idingredient = ingredient.idIngredient;
+                            //if(_model.Ingredient.ToList().FindAll(i => i.Name.ToLower() == ingredient.ingredient.ToLower()).Count() == 0)
+                            //{
+                            //    _model.Ingredient.Add(new Ingredient { Name = ingredient.ingredient });
+                            //    _model.SaveChanges();
+                            //    i.Idingredient = _model.Ingredient.ToList().Last().Idingredient;
+                            //}
+                            //else
+                            //{
+                            //    i.Idingredient = _model.Ingredient.ToList().Find(i => i.Name.ToLower() == ingredient.ingredient.ToLower()).Idingredient;
+                            //}
                             i.Amount = ingredient.amount;
                             _model.IngredientList.Add(i);
                         }
@@ -101,25 +115,34 @@ namespace Backend.Controllers
         /// </summary>
         /// <param name="recipe">Данные рецепта</param>
         /// <returns>Результат, получилось изменить или нет</returns>
-        /// <response code="400">Некорректные значения</response>
         /// <response code="200">Данные рецепта изменены</response>
+        /// <response code="400">Некорректные значения</response>
+        /// <response code="406">Логин пользователя и записанный автор рецепта не совпадает</response>
         /// <response code="500">Внутренняя ошибка (читать сообщение в ответе)</response>
         [HttpPost("api/[controller]/update")]
+        [Authorize(Roles = "admin, open, blocked")]
         public IActionResult Update(FullInfoRecipeBindingModel recipe)
         {
             try
             {
-                if (recipe == null)
+                if (recipe.idRecipe < 1)
                     return BadRequest();
 
-                if (UpdateInDB(recipe))
+                if(CheckAuthorLogin(User.Identity.Name, recipe.idRecipe) || User.IsInRole("admin"))
                 {
-                    return Ok();
+                    var response = UpdateInDB(recipe);
+
+                    if (response == "OK")
+                    {
+                        return Ok();
+                    }
+                    else
+                    {
+                        return StatusCode(500, "Во время изменения что-то пошло не так!" + response);
+                    }
                 }
-                else
-                {
-                    return StatusCode(500, "Во время изменения что-то пошло не так");
-                }
+
+                return StatusCode(406);
             }
             catch (Exception e)
             {
@@ -127,7 +150,7 @@ namespace Backend.Controllers
             }
         }
 
-        private bool UpdateInDB(FullInfoRecipeBindingModel recipe)
+        private string UpdateInDB(FullInfoRecipeBindingModel recipe)
         {
             using (ModelDbContext _model = new ModelDbContext())
             {
@@ -137,16 +160,23 @@ namespace Backend.Controllers
                     {
                         Recipe r = _model.Recipe.ToList().Find(r => r.Idrecipe == recipe.idRecipe);
 
-                        if (recipe.name != null)
+                        if (recipe.name != null && recipe.name.Length > 0)
                             r.Name = recipe.name;
+                        else
+                            throw new Exception("Имя нулевое");
 
                         if (recipe.idIngredient != 0)
                             r.Idingredient = recipe.idIngredient;
+                        else
+                            throw new Exception("Id главного ингредиента присвоен нулю");
+
 
                         if (recipe.idNationality != 0)
                             r.Idnationality = recipe.idNationality;
+                        else
+                            throw new Exception("Id национальности присвоено нулю");
 
-                        if (recipe.duration != null)
+                        if (recipe.duration != null || recipe.duration.Length > 0)
                             r.Duration = TimeSpan.Parse(recipe.duration);
 
                         if (recipe.steps.Count != 0)
@@ -160,6 +190,8 @@ namespace Backend.Controllers
                                 }
                             }
                         }
+                        else
+                            throw new Exception("Отсутствуют шаги");
 
                         if (recipe.ingredientList.Count != 0)
                         {
@@ -178,15 +210,17 @@ namespace Backend.Controllers
                                 }
                             }
                         }
+                        else
+                            throw new Exception("Отсутствует список ингредиентов");
 
                         _model.SaveChanges();
                         transaction.Commit();
-                        return true;
+                        return "OK";
                     }
                     catch (Exception e)
                     {
                         transaction.Rollback();
-                        return false;
+                        return e.Message;
                     }
                 }
             }
@@ -197,10 +231,12 @@ namespace Backend.Controllers
         /// </summary>
         /// <param name="idRecipe">id Рецепта</param>
         /// <returns>Результат, удалён рецепт или нет</returns>
-        /// <response code="400">Некорректные значения</response>
         /// <response code="200">Рецепт удалён</response>
+        /// <response code="400">Некорректные значения</response>
+        /// <response code="406">Логин пользователя и записанный автор рецепта не совпадает</response>
         /// <response code="500">Внутренняя ошибка (читать сообщение в ответе)</response>
         [HttpDelete("api/[controller]/delete")]
+        [Authorize(Roles = "admin, open, blocked")]
         public IActionResult Delete(int idRecipe)
         {
             try
@@ -208,14 +244,20 @@ namespace Backend.Controllers
                 if (idRecipe < 1)
                     return BadRequest();
 
-                if (DeleteFromDB(idRecipe))
+                if (CheckAuthorLogin(User.Identity.Name, idRecipe) || User.IsInRole("admin"))
                 {
-                    return Ok();
+                    if (DeleteFromDB(idRecipe))
+                    {
+                        return Ok();
+                    }
+                    else
+                    {
+                        return StatusCode(500, "Во время удаления что-то пошло не так");
+                    }
                 }
-                else
-                {
-                    return StatusCode(500, "Во время удаления что-то пошло не так");
-                }
+
+                return StatusCode(406);
+                
             }
             catch(Exception e)
             {
@@ -250,6 +292,14 @@ namespace Backend.Controllers
                         return false;
                     }
                 }
+            }
+        }
+
+        private bool CheckAuthorLogin(string login, int idRecipe)
+        {
+            using (ModelDbContext _model = new ModelDbContext())
+            {
+                return login == _model.Recipe.Find(idRecipe).Author;
             }
         }
     }
